+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## v43.x / v44 (In progress)
 - Backgrounds: New flag level 2, color-shift backgrounds. Additionally, very eye-straining backgrounds should be less common at all flag levels.
 - Game improvement patches: "Run button" changed to "Run mode"; default run speed and hold Y to slow to a walk.
+- PC sprites: Added Nessa.
 ### Bugfixes
 - Ancient Cave: Benches no longer play restoration sound effect twice.